change deff_yee2

+++ src/diff_yee2.jl
@@ -7,7 +7,7 @@
 - `kinc`: [kx ky] incident wave vector. This argument is only needed for periodic boundaries.
 Position on the grid: m = (ny - 1)*Nx + nx
 """
-function diff_yee2(NGRID::Matrix{Int64}, RES::Matrix{Float64}, BC::Matrix{Int64}, kinc=nothing)
+function diff_yee2(NGRID::Matrix{Float64}, RES::Matrix{Float64}, BC::Matrix{Int64}, kinc=nothing)
     if kinc != nothing
         kinc0 = kinc
     end
